+++ cli/lib/src/cli_runner.dart
@@ -3,6 +3,7 @@ import 'package:args/command_runner.dart';
 import 'package:mason_logger/mason_logger.dart';
 
 import 'commands/add_command.dart';
+import 'commands/hooks_command.dart';
 import 'commands/install_command.dart';
 import 'commands/quote_command.dart';
 import 'commands/rules_command.dart';
@@ -38,6 +39,7 @@ class SomnioCliRunner extends CommandRunner<int> {
     );
 
     addCommand(AddCommand(logger: _logger));
+    addCommand(HooksCommand(logger: _logger));
     addCommand(InstallCommand(logger: _logger));
     addCommand(RulesCommand(logger: _logger));
     addCommand(UpdateCommand(logger: _logger));